server/requirements.txt
Django==4.0.10
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==5.2.2
django-filters==2.4.0
versatileimagefield==2.0
django-redis==5.4.0
storages==1.14.2
boto3==1.34.72